Component-ized the Line Graph
